material/rest/rest-alunos/requirements.txt
django==4.2
django-ninja==0.22